requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
yfinance==0.2.36
requests==2.31.0
python-dotenv==1.0.0 